change int16 to int8

+++ src/main/drivers/gimbal_common.h
@@ -61,7 +61,7 @@ typedef struct gimbalConfig_s {
     uint8_t panChannel;
     uint8_t tiltChannel;
     uint8_t rollChannel;
-    int16_t sensitivity;
+    int8_t sensitivity;
     int16_t panTrim;
     int16_t tiltTrim;
     int16_t rollTrim;
